Add Azure Key Vault resource identifier and secret client support
