+++ src/Aspire.Hosting.Azure/IKeyVaultResource.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using Aspire.Hosting.ApplicationModel;
+using Azure.Security.KeyVault.Secrets;
 
 namespace Aspire.Hosting.Azure;
 
@@ -21,10 +22,20 @@ public interface IKeyVaultResource : IResource, IAzureResource
     BicepOutputReference NameOutputReference { get; }
 
     /// <summary>
+    /// Gets the resource identifier of the Azure Key Vault resource.
+    /// </summary>
+    BicepOutputReference Id { get; }
+
+    /// <summary>
     /// The secrets for the Azure Key Vault resource. Used in run mode to resolve
     /// the secrets from the Key Vault.
     /// </summary>
     IDictionary<string, string> Secrets { get; }
+
+    /// <summary>
+    /// The secret client used to access the Azure Key Vault at runtime.
+    /// </summary>
+    SecretClient? SecretClient { get; set; }
 
     /// <summary>
     /// Gets a secret reference for the specified secret name.
